+++ src/main/java/com/github/pagehelper/Page.java
@@ -24,6 +24,7 @@
 
 package com.github.pagehelper;
 
+import java.io.Closeable;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -34,7 +35,7 @@ import java.util.List;
  * @version 3.6.0
  *          项目地址 : http://git.oschina.net/free/Mybatis_PageHelper
  */
-public class Page<E> extends ArrayList<E> {
+public class Page<E> extends ArrayList<E> implements Closeable {
     private static final long serialVersionUID = 1L;
 
     /**
@@ -357,5 +358,10 @@ public class Page<E> extends ArrayList<E> {
                 ", reasonable=" + reasonable +
                 ", pageSizeZero=" + pageSizeZero +
                 '}';
+    }
+
+    @Override
+    public void close() {
+        PageHelper.clearPage();
     }
 }