+++ src/test/java/com/github/pagehelper/test/basic/CloseableTest.java
@@ -1,0 +1,54 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014-2017 abel533@gmail.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package com.github.pagehelper.test.basic;
+
+import com.github.pagehelper.mapper.CountryMapper;
+import com.github.pagehelper.model.Country;
+import com.github.pagehelper.util.MybatisHelper;
+import org.apache.ibatis.session.SqlSession;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.List;
+
+public class CloseableTest {
+
+    @Test
+    public void testCloseable() {
+        SqlSession sqlSession = MybatisHelper.getSqlSession();
+        CountryMapper countryMapper = sqlSession.getMapper(CountryMapper.class);
+        //下面注释代码只能在jdk7+中进行测试
+//        try(Page<Object> page = PageHelper.startPage(1, 10)) {
+//            int a = 10/0;
+//            countryMapper.selectAll();
+//            Assert.fail();
+//        } catch (Exception e){
+//            e.printStackTrace();
+//        }
+        List<Country> countries = countryMapper.selectAll();
+        Assert.assertEquals(183, countries.size());
+        sqlSession.close();
+    }
+}